+++ src/main/resources/db/migration/postgresql/V32__task_details.sql
@@ -1,0 +1,2 @@
+ALTER TABLE day_tasks
+    ADD COLUMN description VARCHAR(4000);